# proto/gen/openapi/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/memos:
        get:
            tags:
                - MemoService
            operationId: MemoService_ListMemos
            parameters:
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: query
                  in: query
                  description: |-
                    query, when non-empty, runs a full-text search (FTS5 with a LIKE fallback)
                     instead of returning the recent list.
                  schema:
                    type: string
                - name: archived
                  in: query
                  description: |-
                    archived filters both recent-list and search results before limit is
                     applied. When unset, both current and archived memos are included.
                  schema:
                    type: boolean
                - name: cursor
                  in: query
                  description: |-
                    cursor continues the reverse-chronological recent list. Search requests
                     ignore it because search results are not paginated.
                  schema:
                    type: string
                - name: favorited
                  in: query
                  description: |-
                    favorited filters both recent-list and search results before limit is
                     applied. When unset, both favorited and non-favorited memos are included.
                  schema:
                    type: boolean
                - name: deleted
                  in: query
                  description: |-
                    deleted=true returns recoverable records from Recently Deleted. Purged
                     tombstones are never returned by this user-facing list.
                  schema:
                    type: boolean
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListMemosResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        post:
            tags:
                - MemoService
            operationId: MemoService_CreateMemo
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateMemoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/MemoResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/memos/{id}:
        get:
            tags:
                - MemoService
            operationId: MemoService_GetMemo
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/MemoResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        delete:
            tags:
                - MemoService
            operationId: MemoService_DeleteMemo
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: expectedVersion
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/MemoResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        patch:
            tags:
                - MemoService
            operationId: MemoService_UpdateMemo
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateMemoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/MemoResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
    /api/v1/sync:
        get:
            tags:
                - SyncService
            description: |-
                PullSync returns changes after the given cursor, including tombstones, in
                 pages. Attachments are pull-only (metadata).
            operationId: SyncService_PullSync
            parameters:
                - name: cursor
                  in: query
                  description: |-
                    cursor is an opaque position from a prior response; empty means from the
                     beginning.
                  schema:
                    type: string
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PullSyncResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/sync:push:
        post:
            tags:
                - SyncService
            description: |-
                PushSync applies a batch of client mutations. Each is keyed by mutation_id;
                 a stored prior result can be replayed, while a stale base_version yields a
                 conflict result rather than an error.
            operationId: SyncService_PushSync
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PushSyncRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PushSyncResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
...
        AskConversation:
            type: object
            properties:
                id:
                    type: string
                title:
                    type: string
                status:
                    type: string
                contextScope:
                    type: string
                    description: |-
                        context_scope is the default time window for grounding: "recent_7_days" |
                         "recent_30_days" | "all". Per-message scope/source may override it.
                headMessageId:
                    type: string
                    description: |-
                        head_message_id is the currently active leaf when forks (regenerations /
                         branches) exist.
                pinnedTime:
                    type: string
                    format: date-time
                archivedTime:
                    type: string
                    format: date-time
                createdTime:
                    type: string
                    format: date-time
                updatedTime:
                    type: string
                    format: date-time
                deletedTime:
                    type: string
                    format: date-time
            description: |-
                AskConversation is a single Ask thread (a ChatGPT-style conversation grounded
                 in the user's memos).
...
        AskMessage:
            type: object
            properties:
                id:
                    type: string
                conversationId:
                    type: string
                role:
                    type: string
                    description: role is "user" | "assistant".
                content:
                    type: string
                parentId:
                    type: string
                    description: parent_id is the message this one follows (empty for the first turn).
                forkOfId:
                    type: string
                    description: |-
                        fork_of_id, when set, marks this as an alternative (regenerated) answer to
                         the same parent as the referenced message.
                status:
                    type: string
                    description: |-
                        status is the message lifecycle marker; persisted messages currently use
                         "complete".
                sourceRefs:
                    type: array
                    items:
                        $ref: '#/components/schemas/AskSourceRef'
                model:
                    type: string
                createdTime:
                    type: string
                    format: date-time
                updatedTime:
                    type: string
                    format: date-time
                deletedTime:
                    type: string
                    format: date-time
                promptVersion:
                    type: string
                    description: |-
                        prompt_version identifies the answer instructions used for assistant
                         messages. User and legacy messages leave it empty.
            description: |-
                AskMessage is one turn in a conversation. Branching and regeneration are
                 modeled as a tree: parent_id links a follow-up to the message it answers, and
                 fork_of_id links a regenerated answer to the answer it replaces.
        AskSourceRef:
            type: object
            properties:
                memoId:
                    type: string
                entryDate:
                    type: string
                excerpt:
                    type: string
                rank:
                    type: integer
                    format: int32
            description: |-
                AskSourceRef is a citation: a memo an answer was grounded in, so the user can
                 jump back to the original record.
        Attachment:
            type: object
            properties:
                id:
                    type: string
                uid:
                    type: string
                memoId:
                    type: string
                url:
                    type: string
                filename:
                    type: string
                contentType:
                    type: string
                size:
                    type: string
                sha256:
                    type: string
                width:
                    type: string
                height:
                    type: string
                status:
                    type: string
                    description: status is the lifecycle marker; stored local attachments use "stored".
                createdTime:
                    type: string
                    format: date-time
                updatedTime:
                    type: string
                    format: date-time
                deletedTime:
                    type: string
                    format: date-time
            description: |-
                Attachment is a file stored alongside a memo. Bytes live on disk; this
                 message carries only metadata. In sync, attachments are pull-only (the
                 server never accepts attachment pushes).
...
        CreateMemoRequest:
            type: object
            properties:
                id:
                    type: string
                    description: |-
                        id may be client-supplied so an offline memo keeps the same identity after
                         sync. A duplicate id is not itself an idempotency key.
                content:
                    type: string
                entryDate:
                    type: string
...
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        ListMemosResponse:
            type: object
            properties:
                memos:
                    type: array
                    items:
                        $ref: '#/components/schemas/Memo'
                nextCursor:
                    type: string
                    description: next_cursor is empty when the recent list is exhausted and for searches.
...
        Memo:
            type: object
            properties:
                id:
                    type: string
                content:
                    type: string
                entryDate:
                    type: string
                    description: |-
                        entry_date is the user-chosen calendar day in YYYY-MM-DD form; it is
                         independent of created_time.
                version:
                    type: string
                    description: |-
                        version is bumped on every content/state mutation and drives optimistic
                         concurrency: writes carry the expected_version they read.
                archivedTime:
                    type: string
                    format: date-time
                createdTime:
                    type: string
                    format: date-time
                updatedTime:
                    type: string
                    format: date-time
                deletedTime:
                    type: string
                    description: deleted_time set => the memo is tombstoned (soft-deleted).
                    format: date-time
                favoritedAt:
                    type: string
                    format: date-time
                purgedAt:
                    type: string
                    format: date-time
            description: |-
                Memo is Sillage's only content unit; user-facing clients present it as a record.
                 Deletion first creates a recoverable tombstone. After permanent purge,
                 purged_time is set and user content has been scrubbed while the minimal row
                 remains so sync clients can converge.
        MemoAI:
            type: object
            properties:
                memoId:
                    type: string
                summary:
                    type: string
                sentiment:
                    type: string
                provider:
                    type: string
                model:
                    type: string
                profileId:
                    type: string
                promptVersion:
                    type: string
                sourceMemoIds:
                    type: string
                    description: source_memo_ids is a JSON array of the memo IDs the summary is grounded in.
                status:
                    type: string
                    description: |-
                        status is the generation lifecycle marker; successful summaries use
                         "complete".
                errorCode:
                    type: string
                    description: error_code explains a "failed"/degraded status, e.g. "key_unavailable".
                startedTime:
                    type: string
                    format: date-time
                finishedTime:
                    type: string
                    format: date-time
                inputTokens:
                    type: string
                outputTokens:
                    type: string
                totalTokens:
                    type: string
                createdTime:
                    type: string
                    format: date-time
                updatedTime:
                    type: string
                    format: date-time
            description: |-
                MemoAI holds AI-derived data for a memo (currently the summary). It is stored
                 in a separate table and intentionally does NOT bump the memo's version or
                 updated_time, so regenerating a summary never looks like a content edit.
        MemoResponse:
            type: object
            properties:
                memo:
                    $ref: '#/components/schemas/Memo'
                ai:
                    allOf:
                        - $ref: '#/components/schemas/MemoAI'
                    description: |-
                        ai carries the memo's latest summary when one exists, so clients can render
                         a stored summary without regenerating it.
...
        PullSyncResponse:
            type: object
            properties:
                memos:
                    type: array
                    items:
                        $ref: '#/components/schemas/Memo'
                attachments:
                    type: array
                    items:
                        $ref: '#/components/schemas/Attachment'
                memoAi:
                    type: array
                    items:
                        $ref: '#/components/schemas/MemoAI'
                askConversations:
                    type: array
                    items:
                        $ref: '#/components/schemas/AskConversation'
                askMessages:
                    type: array
                    items:
                        $ref: '#/components/schemas/AskMessage'
                cursor:
                    type: string
                    description: |-
                        cursor and next_cursor are the server position after this page. Clients
                         should persist next_cursor only after applying the returned resources.
                nextCursor:
                    type: string
                hasMore:
                    type: boolean
...
        PushSyncRequest:
            type: object
            properties:
                changes:
                    type: array
                    items:
                        $ref: '#/components/schemas/SyncChange'
        PushSyncResponse:
            type: object
            properties:
                results:
                    type: array
                    items:
                        $ref: '#/components/schemas/SyncResult'
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
        SyncChange:
            type: object
            properties:
                mutationId:
                    type: string
                    description: mutation_id is a client-generated unique key used to find a stored result.
                resourceType:
                    type: string
                resourceId:
                    type: string
                action:
                    type: string
                    description: action is "create" | "update" | "delete" | "restore" | "purge".
                baseVersion:
                    type: string
                    description: |-
                        base_version is the memo version the client based this change on; used for
                         conflict detection on update/delete/restore/purge. Create omits it.
                localChangedAt:
                    type: string
                memo:
                    $ref: '#/components/schemas/SyncMemoPayload'
            description: |-
                SyncChange is one client mutation. Only resource_type "memo" is accepted on
                 push today.
...
        UpdateMemoRequest:
            type: object
            properties:
                id:
                    type: string
                expectedVersion:
                    type: string
                content:
                    type: string
                entryDate:
                    type: string
                archived:
                    type: boolean
                favorited:
                    type: boolean
            description: |-
                UpdateMemo applies a partial update guarded by expected_version. Unset
                 optional fields are left unchanged.
...