# server/memo_routes.go
package server

import (
	"database/sql"
	"errors"
	"net/http"
	"strconv"
	"strings"
	"time"

	"github.com/labstack/echo/v5"

	memoapp "github.com/getsillage/sillage/server/memo"
	"github.com/getsillage/sillage/store"
)

type memoRequest struct {
	ID              string  `json:"id"`
	Content         *string `json:"content"`
	EntryDate       *string `json:"entryDate"`
	ExpectedVersion int64   `json:"expectedVersion"`
	Favorited       *bool   `json:"favorited"`
	Archived        *bool   `json:"archived"`
}

func (s *Server) registerMemoRoutes(e *echo.Echo) {
	e.GET("/api/v1/memos", s.handleListMemos)
	e.POST("/api/v1/memos", s.handleCreateMemo)
	e.GET("/api/v1/memos/:memo", s.handleGetMemo)
	e.PATCH("/api/v1/memos/:memo", s.handleUpdateMemo)
	e.DELETE("/api/v1/memos/:memo", s.handleDeleteMemo)
	e.POST("/api/v1/memos/:memo", s.handleMemoAction)
	e.GET("/api/v1/sync", s.handleSyncPull)
	e.POST("/api/v1/sync:push", s.handleSyncPush)
}

func (s *Server) handleListMemos(c *echo.Context) error {
	account, err := s.accountFromBearer(c)
	if err != nil {
		return apiError(c, http.StatusUnauthorized, "unauthenticated", "请重新登录")
	}
	limit := parseLimit(c.QueryParam("limit"), 50)
	ctx := c.Request().Context()
	// Accept both "query" (canonical, matches proto) and legacy "q".
	query := c.QueryParam("query")
	if query == "" {
		query = c.QueryParam("q")
	}
	archived, err := parseMemoBoolFilter(c.QueryParam("archived"), "archived")
	if err != nil {
		return apiError(c, http.StatusBadRequest, "invalid_field", err.Error())
	}
	favorited, err := parseMemoBoolFilter(c.QueryParam("favorited"), "favorited")
	if err != nil {
		return apiError(c, http.StatusBadRequest, "invalid_field", err.Error())
	}
	deleted, err := parseMemoBoolFilter(c.QueryParam("deleted"), "deleted")
	if err != nil {
		return apiError(c, http.StatusBadRequest, "invalid_field", err.Error())
	}
	if query != "" {
		memos, err := s.memos.Search(ctx, account.ID, memoapp.SearchInput{
			Query:     query,
			Archived:  archived,
			Favorited: favorited,
			Deleted:   deleted,
			Limit:     limit,
		})
		if err != nil {
			if isValidationError(err) {
				return apiError(c, http.StatusBadRequest, "invalid_field", err.Error())
			}
			return apiError(c, http.StatusInternalServerError, "internal", "读取记录失败")
		}
		return c.JSON(http.StatusOK, map[string]any{"memos": memoDTOs(memos)})
	}
	page, err := s.memos.List(ctx, account.ID, memoapp.ListInput{
		Archived:  archived,
		Favorited: favorited,
		Deleted:   deleted,
		Limit:     limit,
		Cursor:    c.QueryParam("cursor"),
	})
	if err != nil {
		return apiError(c, http.StatusInternalServerError, "internal", "读取记录失败")
	}
	return c.JSON(http.StatusOK, map[string]any{
		"memos":      memoDTOs(page.Memos),
		"nextCursor": page.NextCursor,
	})
}

func (s *Server) handleCreateMemo(c *echo.Context) error {
	account, err := s.accountFromBearer(c)
	if err != nil {
		return apiError(c, http.StatusUnauthorized, "unauthenticated", "请重新登录")
	}
	var req memoRequest
	if err := c.Bind(&req); err != nil {
		return apiError(c, http.StatusBadRequest, "invalid_json", "请求格式不正确")
	}
	content := stringValue(req.Content)
	entryDate := stringValue(req.EntryDate)
	memo, err := s.memos.Create(c.Request().Context(), account.ID, memoapp.CreateInput{
		ID:        req.ID,
		Content:   content,
		EntryDate: entryDate,
	})
	if err != nil {
		if isValidationError(err) {
			return apiError(c, http.StatusBadRequest, "invalid_field", err.Error())
		}
		return apiError(c, http.StatusInternalServerError, "internal", "保存记录失败")
	}
	return c.JSON(http.StatusOK, map[string]any{"memo": memoDTO(memo)})
}

func (s *Server) handleGetMemo(c *echo.Context) error {
	account, err := s.accountFromBearer(c)
	if err != nil {
		return apiError(c, http.StatusUnauthorized, "unauthenticated", "请重新登录")
	}
	ctx := c.Request().Context()
	detail, err := s.memos.Get(ctx, account.ID, c.Param("memo"))
	if err != nil {
		if errors.Is(err, memoapp.ErrSummaryRead) {
			return apiError(c, http.StatusInternalServerError, "internal", "读取总结失败")
		}
		if errors.Is(err, sql.ErrNoRows) {
			return apiError(c, http.StatusNotFound, "not_found", "记录不存在")
		}
		return apiError(c, http.StatusInternalServerError, "internal", "读取记录失败")
	}
	body := map[string]any{"memo": memoDTO(detail.Memo)}
	if detail.AI != nil {
		body["ai"] = memoAIDTO(detail.AI)
	}
	return c.JSON(http.StatusOK, body)
}

func (s *Server) handleUpdateMemo(c *echo.Context) error {
	account, err := s.accountFromBearer(c)
	if err != nil {
		return apiError(c, http.StatusUnauthorized, "unauthenticated", "请重新登录")
	}
	var req memoRequest
	if err := c.Bind(&req); err != nil {
		return apiError(c, http.StatusBadRequest, "invalid_json", "请求格式不正确")
	}
	memo, err := s.memos.Update(c.Request().Context(), account.ID, memoapp.UpdateInput{
		ID:              memoParam(c),
		ExpectedVersion: req.ExpectedVersion,
		Content:         req.Content,
		EntryDate:       req.EntryDate,
		Favorited:       req.Favorited,
		Archived:        req.Archived,
	})
	return s.writeMemoMutationResult(c, memo, err)
}

func (s *Server) handleDeleteMemo(c *echo.Context) error {
	account, err := s.accountFromBearer(c)
	if err != nil {
		return apiError(c, http.StatusUnauthorized, "unauthenticated", "请重新登录")
	}
	expectedVersion, err := parseExpectedVersionParam(c)
	if err != nil {
		return apiError(c, http.StatusBadRequest, "invalid_field", err.Error())
	}
	memo, err := s.memos.Delete(c.Request().Context(), account.ID, memoParam(c), expectedVersion)
	return s.writeMemoMutationResult(c, memo, err)
}

func (s *Server) handleMemoAction(c *echo.Context) error {
	account, err := s.accountFromBearer(c)
	if err != nil {
		return apiError(c, http.StatusUnauthorized, "unauthenticated", "请重新登录")
	}
	memoID, action, ok := memoActionParam(c)
	if !ok {
		return apiError(c, http.StatusNotFound, "not_found", "接口不存在")
	}

	if action == "generate-summary" {
		ai, err := s.generateMemoSummary(c.Request().Context(), account.ID, memoID)
		if err != nil {
			status, code, message := memoHTTPStatus(err)
			if code == "internal" {
				message = "生成总结失败"
			}
			return apiError(c, status, code, message)
		}
		return c.JSON(http.StatusOK, map[string]any{"ai": memoAIDTO(ai)})
	}

	update := memoapp.UpdateInput{ID: memoID}
	switch action {
	case "restore", "purge":
		var req memoRequest
		if err := c.Bind(&req); err != nil {
			return apiError(c, http.StatusBadRequest, "invalid_json", "请求格式不正确")
		}
		var memo *store.Memo
		if action == "restore" {
			memo, err = s.memos.Restore(c.Request().Context(), account.ID, memoID, req.ExpectedVersion)
		} else {
			memo, err = s.memos.Purge(c.Request().Context(), account.ID, memoID, req.ExpectedVersion)
		}
		return s.writeMemoMutationResult(c, memo, err)
	case "setArchived":
		var req memoRequest
		if err := c.Bind(&req); err != nil {
			return apiError(c, http.StatusBadRequest, "invalid_json", "请求格式不正确")
		}
		if req.Archived == nil {
			return apiError(c, http.StatusBadRequest, "invalid_field", "archived 必须是 true 或 false")
		}
		value := *req.Archived
		update.ExpectedVersion = req.ExpectedVersion
		update.Archived = &value
	case "setFavorited":
		var req memoRequest
		if err := c.Bind(&req); err != nil {
			return apiError(c, http.StatusBadRequest, "invalid_json", "请求格式不正确")
		}
		if req.Favorited == nil {
			return apiError(c, http.StatusBadRequest, "invalid_field", "favorited 必须是 true 或 false")
		}
		value := *req.Favorited
		update.ExpectedVersion = req.ExpectedVersion
		update.Favorited = &value
	case "archive":
		value := true
		update.ExpectedVersion, err = parseExpectedVersionParam(c)
		if err != nil {
			return apiError(c, http.StatusBadRequest, "invalid_field", err.Error())
		}
		update.Archived = &value
	case "unarchive":
		value := false
		update.ExpectedVersion, err = parseExpectedVersionParam(c)
		if err != nil {
			return apiError(c, http.StatusBadRequest, "invalid_field", err.Error())
		}
		update.Archived = &value
	default:
		return apiError(c, http.StatusNotFound, "not_found", "接口不存在")
	}
	memo, err := s.memos.Update(c.Request().Context(), account.ID, update)
	return s.writeMemoMutationResult(c, memo, err)
}

func (s *Server) writeMemoMutationResult(c *echo.Context, memo *store.Memo, err error) error {
	if err == nil {
		return c.JSON(http.StatusOK, map[string]any{"memo": memoDTO(memo)})
	}
	var conflict *store.MemoConflictError
	switch {
	case errors.As(err, &conflict):
		return c.JSON(http.StatusConflict, map[string]any{
			"error": map[string]any{
				"code":          "version_conflict",
				"message":       "记录已被其他修改更新",
				"serverMemo":    memoDTO(conflict.ServerMemo),
				"serverVersion": conflict.ServerMemo.Version,
			},
		})
	case errors.Is(err, sql.ErrNoRows):
		return apiError(c, http.StatusNotFound, "not_found", "记录不存在")
	case isValidationError(err):
		return apiError(c, http.StatusBadRequest, "invalid_field", err.Error())
	default:
		return apiError(c, http.StatusInternalServerError, "internal", "保存记录失败")
	}
}

func memoParam(c *echo.Context) string {
	if value := c.Param("memo"); value != "" {
		return value
	}
	path := c.Request().URL.Path
	prefix := "/api/v1/memos/"
	if !strings.HasPrefix(path, prefix) {
		return ""
	}
	value := strings.TrimPrefix(path, prefix)
	if before, _, ok := strings.Cut(value, ":"); ok {
		return before
	}
	if before, _, ok := strings.Cut(value, "/"); ok {
		return before
	}
	return value
}

func memoActionParam(c *echo.Context) (string, string, bool) {
	value := c.Param("memo")
	if value == "" {
		path := c.Request().URL.Path
		prefix := "/api/v1/memos/"
		if !strings.HasPrefix(path, prefix) {
			return "", "", false
		}
		value = strings.TrimPrefix(path, prefix)
	}
	memoID, action, ok := strings.Cut(value, ":")
	if !ok || memoID == "" || action == "" {
		return "", "", false
	}
	return memoID, action, true
}

func parseLimit(raw string, fallback int) int {
	limit, err := strconv.Atoi(raw)
	if err != nil || limit <= 0 {
		return fallback
	}
	return normalizeLimit(limit, fallback)
}

func parseMemoBoolFilter(raw, field string) (*bool, error) {
	if raw == "" {
		return nil, nil
	}
	value, err := strconv.ParseBool(raw)
	if err != nil {
		return nil, errors.New(field + " 必须是 true 或 false")
	}
	return &value, nil
}

// parseExpectedVersionParam rejects malformed values outright so a typo'd
// query parameter fails as invalid_field instead of the misleading
// "expectedVersion 必须大于 0" validation error downstream.
func parseExpectedVersionParam(c *echo.Context) (int64, error) {
	raw := c.QueryParam("expectedVersion")
	if raw == "" {
		return 0, nil
	}
	value, err := strconv.ParseInt(raw, 10, 64)
	if err != nil {
		return 0, errors.New("expectedVersion 必须是数字")
	}
	return value, nil
}

func stringValue(value *string) string {
	if value == nil {
		return ""
	}
	return *value
}

func memoDTOs(memos []*store.Memo) []map[string]any {
	dtos := make([]map[string]any, 0, len(memos))
	for _, memo := range memos {
		dtos = append(dtos, memoDTO(memo))
	}
	return dtos
}

func memoDTO(memo *store.Memo) map[string]any {
	if memo == nil {
		return nil
	}
	return map[string]any{
		"id":          memo.ID,
		"content":     memo.Content,
		"entryDate":   memo.EntryDate,
		"version":     memo.Version,
		"favoritedAt": optionalTime(memo.FavoritedAt),
		"archivedAt":  optionalTime(memo.ArchivedAt),
		"createdAt":   time.UnixMilli(memo.CreatedAt).UTC().Format(time.RFC3339),
		"updatedAt":   time.UnixMilli(memo.UpdatedAt).UTC().Format(time.RFC3339),
		"deletedAt":   optionalTime(memo.DeletedAt),
		"purgedAt":    optionalTime(memo.PurgedAt),
	}
}

func optionalTime(value sql.NullInt64) any {
	if !value.Valid {
		return nil
	}
	return time.UnixMilli(value.Int64).UTC().Format(time.RFC3339)
}